refactor: main 태그의 1개의 자식 컴포넌트를 Filter, RestaurantList로 분리
